requirements.txt
ipykernel
langchain
langchain-core
langchain-openai
python-dotenv
pydantic
defusedxml